requirements.txt
langchain>=0.3
langchain-google-genai>=2.0
google-generativeai>=0.8
streamlit>=1.38
pandas>=2.0
scikit-learn>=1.3
matplotlib>=3.8
seaborn>=0.13
plotly>=5.18
python-dotenv>=1.0
scipy>=1.11
tqdm>=4.66
